LLM_Client/requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.7
redis==5.0.1
